+++ testkit-backend/src/main/java/neo4j/org/testkit/backend/channel/handler/TestkitRequestResponseMapperHandler.java
@@ -19,26 +19,17 @@
 package neo4j.org.testkit.backend.channel.handler;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.netty.channel.ChannelDuplexHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelPromise;
-import neo4j.org.testkit.backend.messages.TestkitModule;
+import neo4j.org.testkit.backend.CommandProcessor;
 import neo4j.org.testkit.backend.messages.requests.TestkitRequest;
 import neo4j.org.testkit.backend.messages.responses.TestkitResponse;
 
 public class TestkitRequestResponseMapperHandler extends ChannelDuplexHandler
 {
-    private final ObjectMapper objectMapper;
-
-    public TestkitRequestResponseMapperHandler()
-    {
-        objectMapper = new ObjectMapper();
-        TestkitModule testkitModule = new TestkitModule();
-        this.objectMapper.registerModule( testkitModule );
-        this.objectMapper.disable( DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES );
-    }
+    private final ObjectMapper objectMapper = CommandProcessor.newObjectMapperFor( () -> true );
 
     @Override
     public void channelRead( ChannelHandlerContext ctx, Object msg )
