+++ testkit-backend/src/main/java/neo4j/org/testkit/backend/messages/requests/SessionReadTransaction.java
@@ -18,10 +18,13 @@
  */
 package neo4j.org.testkit.backend.messages.requests;
 
+import com.fasterxml.jackson.annotation.JacksonInject;
+import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import neo4j.org.testkit.backend.AsyncSessionState;
+import neo4j.org.testkit.backend.CommandProcessor;
 import neo4j.org.testkit.backend.SessionState;
 import neo4j.org.testkit.backend.TestkitState;
 import neo4j.org.testkit.backend.messages.responses.RetryableDone;
@@ -39,10 +42,16 @@ import org.neo4j.driver.async.AsyncTransactionWork;
 
 @Setter
 @Getter
-@NoArgsConstructor
 public class SessionReadTransaction implements TestkitRequest
 {
+    private final CommandProcessor commandProcessor;
+
     private SessionReadTransactionBody data;
+
+    public SessionReadTransaction( @JacksonInject(CommandProcessor.COMMAND_PROCESSOR_ID) CommandProcessor commandProcessor )
+    {
+        this.commandProcessor = commandProcessor;
+    }
 
     @Override
     public TestkitResponse process( TestkitState testkitState )
@@ -90,7 +99,7 @@ public class SessionReadTransaction implements TestkitRequest
             while ( true )
             {
                 // Process commands as usual but blocking in here
-                testkitState.getProcessor().get();
+                commandProcessor.process();
 
                 // Check if state changed on session
                 switch ( sessionState.retryableState )
