+++ testkit-backend/src/test/java/neo4j/org/testkit/backend/MessageDeserializerTest.java
@@ -19,14 +19,11 @@
 package neo4j.org.testkit.backend;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import neo4j.org.testkit.backend.messages.TestkitModule;
 import neo4j.org.testkit.backend.messages.requests.NewDriver;
 import neo4j.org.testkit.backend.messages.requests.NewSession;
-import neo4j.org.testkit.backend.messages.requests.TestkitRequest;
 import neo4j.org.testkit.backend.messages.requests.SessionRun;
-import org.junit.jupiter.api.BeforeAll;
+import neo4j.org.testkit.backend.messages.requests.TestkitRequest;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -35,16 +32,7 @@ import static org.hamcrest.Matchers.instanceOf;
 
 class MessageDeserializerTest
 {
-    private static final ObjectMapper mapper = new ObjectMapper();
-
-    @BeforeAll
-    static void setUp()
-    {
-        TestkitModule tkm = new TestkitModule();
-
-        //mapper.registerModule( tkm );
-        mapper.disable( DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES );
-    }
+    private static final ObjectMapper mapper = CommandProcessor.newObjectMapperFor( () -> false );
 
     @Test
     void testDeserializeNewDriver() throws JsonProcessingException
