+++ testkit-backend/src/test/java/neo4j/org/testkit/backend/MessageSerializerTest.java
@@ -19,11 +19,8 @@
 package neo4j.org.testkit.backend;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import neo4j.org.testkit.backend.messages.TestkitModule;
 import neo4j.org.testkit.backend.messages.responses.Driver;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -31,16 +28,7 @@ import static org.hamcrest.Matchers.equalTo;
 
 public class MessageSerializerTest
 {
-    private static final ObjectMapper mapper = new ObjectMapper();
-
-    @BeforeAll
-    static void setUp()
-    {
-        TestkitModule tkm = new TestkitModule();
-
-        mapper.registerModule( tkm );
-        mapper.disable( DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES );
-    }
+    private static final ObjectMapper mapper = CommandProcessor.newObjectMapperFor( () -> false );
 
     @Test
     void shouldSerializerNewDriverResponse() throws JsonProcessingException
